Search and Follower/Following fixed
In this commit, a "whichUser" function was created that checks if the localStorage's user handle is the same as the one being passed in is the same. If so, the user get's routed to the user profile or else it'll be routed to view profile. Search functionality was also implemented.